fix: Prevent premature init of org-roam
helm-org-rifle and org-journal were triggering the loading of org and
org-roam before they had a chance to be initialized. This was only a
problem for org-roam: it would crash the loading of emacs config when
~/org/roam did not exist. So I had to move these packages for after
org-roam in config.el, breaking the alphabetic order. Alternatively, I
could have just created an empty ~/org/roam directory.