Uniform behaviour for web snippets
Behaviour of org-protocol when clicking the browser extension: open the
capture window or dump it directly to the inbox?

It does not make sense to have one behaviour for getting just the URL
and another for getting the URL plus quote. I'm going to just dump this
directly into inbox.org and then have it open on the side if I want to
paste the snippet into the note I'm currently working on.