Adapt config to new Org structure by client
The Org files now have optional directories per client. This was
necessary to keep DBG files out of git and sync'ed through the company's
OneDrive. The new structure is also easy to reason with in general,
although it requires changing the doom config when adding/removing
clients. This, however, happens very rarely, so it's OK.